Expand style, html structure
Filled out html and css to style the calculator.
Added a styling for on button press. I'd like
there to be two different functionalities:
1. Keystroke, and 2. button click. Both
should have the same effect on style
and script.

Learned a little about animations for the
blinker input.

Flexbox for keyboard was a little complicated,
but pretty happy with how it came out.