Expand calc functionality in chaining operations
Calculator functionaly now allows for chaining operations in an
effective and coherent fashion as detailed in calculator-events.js.
Rudimentary unit tests have been added and will be expanded later.

I now need to retool the DOM to respond to these events and do
what I want in terms of display based on certain key strikes.

For example, on "enter", I want the DOM to render
calculator.result() in the input bar, and on the
subsequent keystrike, render the new input.